undo some package moves
