+++ internal/api/freight.go
@@ -12,8 +12,8 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	kargoapi "github.com/akuity/kargo/api/v1alpha1"
+	"github.com/akuity/kargo/internal/git"
 	"github.com/akuity/kargo/internal/helm"
-	"github.com/akuity/kargo/pkg/git"
 )
 
 // GenerateFreightID deterministically calculates a piece of Freight's ID based
