+++ internal/controller/freight/finder.go
@@ -8,8 +8,8 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	kargoapi "github.com/akuity/kargo/api/v1alpha1"
-	libGit "github.com/akuity/kargo/pkg/git"
-	pkgwarehouse "github.com/akuity/kargo/pkg/warehouse"
+	"github.com/akuity/kargo/internal/api"
+	libGit "github.com/akuity/kargo/internal/git"
 )
 
 type NotFoundError struct {
@@ -37,7 +37,7 @@ func FindCommit(
 	if desiredOrigin == nil {
 		for i := range freightReqs {
 			requestedFreight := freightReqs[i]
-			warehouse, err := pkgwarehouse.GetWarehouse(
+			warehouse, err := api.GetWarehouse(
 				ctx,
 				cl,
 				types.NamespacedName{
@@ -110,7 +110,7 @@ func FindImage(
 	if desiredOrigin == nil {
 		for i := range freightReqs {
 			requestedFreight := freightReqs[i]
-			warehouse, err := pkgwarehouse.GetWarehouse(
+			warehouse, err := api.GetWarehouse(
 				ctx,
 				cl,
 				types.NamespacedName{
@@ -174,7 +174,7 @@ func HasAmbiguousImageRequest(
 
 	for i := range freightReqs {
 		requestedFreight := freightReqs[i]
-		warehouse, err := pkgwarehouse.GetWarehouse(
+		warehouse, err := api.GetWarehouse(
 			ctx,
 			cl,
 			types.NamespacedName{
@@ -226,7 +226,7 @@ func FindChart(
 	if desiredOrigin == nil {
 		for i := range freightReqs {
 			requestedFreight := freightReqs[i]
-			warehouse, err := pkgwarehouse.GetWarehouse(
+			warehouse, err := api.GetWarehouse(
 				ctx,
 				cl,
 				types.NamespacedName{
