+++ internal/controller/warehouses/warehouses.go
@@ -22,7 +22,6 @@ import (
 	"github.com/akuity/kargo/internal/kubeclient"
 	"github.com/akuity/kargo/internal/logging"
 	intpredicate "github.com/akuity/kargo/internal/predicate"
-	pkgwarehouse "github.com/akuity/kargo/pkg/warehouse"
 )
 
 type ReconcilerConfig struct {
@@ -138,7 +137,7 @@ func (r *reconciler) Reconcile(
 	logger.Debug("reconciling Warehouse")
 
 	// Find the Warehouse
-	warehouse, err := pkgwarehouse.GetWarehouse(ctx, r.client, req.NamespacedName)
+	warehouse, err := api.GetWarehouse(ctx, r.client, req.NamespacedName)
 	if err != nil {
 		return ctrl.Result{}, err
 	}
