+++ internal/credentials/kubernetes/database.go
@@ -17,10 +17,10 @@ import (
 	"github.com/akuity/kargo/internal/credentials/kubernetes/ecr"
 	"github.com/akuity/kargo/internal/credentials/kubernetes/gar"
 	"github.com/akuity/kargo/internal/credentials/kubernetes/github"
+	"github.com/akuity/kargo/internal/git"
 	"github.com/akuity/kargo/internal/helm"
 	"github.com/akuity/kargo/internal/image"
 	"github.com/akuity/kargo/internal/logging"
-	"github.com/akuity/kargo/pkg/git"
 )
 
 // database is an implementation of the credentials.Database interface that
