+++ internal/gitprovider/gitea/gitea.go
@@ -10,8 +10,8 @@ import (
 	"code.gitea.io/sdk/gitea"
 	"github.com/hashicorp/go-cleanhttp"
 
+	"github.com/akuity/kargo/internal/git"
 	"github.com/akuity/kargo/internal/gitprovider"
-	"github.com/akuity/kargo/pkg/git"
 )
 
 const ProviderName = "gitea"
