+++ internal/promotion/runner/builtin/kustomize_image_setter.go
@@ -19,8 +19,8 @@ import (
 	kyaml "sigs.k8s.io/kustomize/kyaml/yaml"
 
 	kargoapi "github.com/akuity/kargo/api/v1alpha1"
+	"github.com/akuity/kargo/internal/controller/freight"
 	intyaml "github.com/akuity/kargo/internal/yaml"
-	"github.com/akuity/kargo/pkg/freight"
 	"github.com/akuity/kargo/pkg/promotion"
 	"github.com/akuity/kargo/pkg/x/promotion/runner/builtin"
 )
