+++ internal/webhook/external/refresh_test.go
@@ -15,8 +15,8 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client/interceptor"
 
 	kargoapi "github.com/akuity/kargo/api/v1alpha1"
+	"github.com/akuity/kargo/internal/git"
 	"github.com/akuity/kargo/internal/indexer"
-	"github.com/akuity/kargo/pkg/git"
 )
 
 func TestRefreshWarehouses(t *testing.T) {
