--- pkg/freight/doc.go
@@ -1,2 +1,0 @@
-// Package freight provides functions for finding and resolving resources from Freight objects.
-package freight
