--- pkg/git/doc.go
@@ -1,2 +1,0 @@
-// Package git provides functions for normalizing and managing git URLs
-package git
