--- pkg/warehouse/doc.go
@@ -1,2 +1,0 @@
-// Package warehouse provides methods for fetching and refreshing warehouse resources from the Kargo API
-package warehouse
